Create Date
Problem Statement: Programmers often develop a "kit bag" of useful classes that they find they
need for almost every project. Here is your chance to begin your own collection and learn C++. For
this exercise, you are to design and implement a simple class whose purpose is to represent a date.
Please come to the tutorial with the methods and data members of your class. Make sure
your Date class has the following abilities:
 To print the date to a stream (e.g. myDate.print(ostream& s))
 To return the date as a text string
 To set myself to the current date
 To set the day, month, and year (individually)
 To find out the day of the week
 To find out the day of the year
 To find out if the year is a leap year
 To find out the week of the year
Hints:
1. Keep your internal representation private and use accessor methods to get and set
components of the date.
2. A nice way to handle things global to a class is to use static class methods and data
members.